refactor(pricing): split stay pricing into rules, reads, and orchestration
`domain/booking/pricing.rs` was 607 lines doing three jobs at once: the
pure pricing rules, ten SQL loaders, and the orchestrators that tie them
together. It was the only non-trivial file in `domain/` and the only one
issuing SQL, which is why step 1's architecture guard had to carve out an
exception for it.

Split it along those seams:

  domain/booking/pricing.rs        217 lines, pure rules, 0 sqlx refs
  queries/booking/pricing_queries.rs  313 lines, the reads
  services/booking/pricing_service.rs  25 lines, load-then-calculate

Its private `read_f64` and `read_money_vnd` were byte-identical copies of
the helpers consolidated into `db::row` in the previous commit; the query
module now uses those instead.

Drop the guard exception and the meta-guard that watched it, since
`the_domain_layer_does_not_talk_to_sqlite` now passes with no carve-out.

Delete the dead pool-based pricing path. With the code no longer hidden
behind `#[allow(dead_code)]`, clippy reported six never-used functions:
production always prices inside its own transaction so it can read rows
it has written but not committed, so the pool variants were reachable
only from tests. Every assertion survives — the pool-vs-tx comparison
test becomes a tx test keeping all its literal expectations (and gains an
explicit `pricing_type` assertion it previously only compared), and the
two deleted error-path tests were verified byte-identical to their
existing `_tx` counterparts first.

No behavior change. cargo test 847 -> 844 (-1 obsolete meta-guard,
-2 duplicate error-path tests); clippy clean; npm test 358 passed;
`npm run verify:full` exit 0 including the native app smoke boot.
