+++ mhm/src-tauri/src/queries/booking/mod.rs
@@ -1,5 +1,6 @@
 pub mod audit_queries;
 pub mod billing_queries;
 pub mod ceo_read_queries;
+pub mod pricing_queries;
 pub mod revenue_queries;
 pub mod room_queries;
