+++ mhm/src-tauri/src/services/booking/mod.rs
@@ -4,6 +4,7 @@ pub mod group_lifecycle;
 pub mod group_service_management;
 pub mod guest_service;
 pub mod invoice_generation;
+pub mod pricing_service;
 pub mod reservation_lifecycle;
 pub mod stay_lifecycle;
 pub mod support;
