+++ mhm/src-tauri/src/services/booking/reservation_lifecycle.rs
@@ -12,9 +12,7 @@ use crate::{
         WriteCommandExecutor, WriteCommandRequest,
     },
     db_error_monitoring::{classify_db_error_code, is_room_unavailable_conflict_message},
-    domain::booking::{
-        pricing::calculate_stay_price_tx, BookingError, BookingResult, OriginSideEffect,
-    },
+    domain::booking::{BookingError, BookingResult, OriginSideEffect},
     models::{status, Booking, CreateReservationRequest, ModifyReservationRequest},
     money::{validate_non_negative_money_vnd, MoneyVnd},
     outbox::{OutboxAggregateKeySource, OutboxEventSpec},
@@ -26,6 +24,7 @@ use super::{
         record_charge_with_origin_tx, record_deposit_tx, record_deposit_with_origin_tx,
     },
     guest_service::{create_reservation_guest_manifest, link_booking_guests},
+    pricing_service::calculate_stay_price_tx,
     support::{
         ensure_one_row_affected, insert_room_calendar_rows, invalid_state_transition,
         read_money_vnd_or_zero, read_money_vnd_strict, validate_non_negative_booking_money,
